--- a/composer.json
+++ b/composer.json
@@ -12,7 +12,8 @@
     },
     "require-dev": {
         "friendsofphp/php-cs-fixer": "^1.13.1",
-        "phpunit/phpunit": "^5.7.4"
+        "phpunit/phpunit": "^5.7.4",
+        "phpbench/phpbench": "^0.13.0"
     },
     "autoload": {
         "psr-4": {
@@ -24,7 +25,8 @@
     },
     "autoload-dev": {
         "psr-4": {
-            "Roave\\BetterReflectionTest\\": "test/unit"
+            "Roave\\BetterReflectionTest\\": "test/unit",
+            "Roave\\BetterReflectionBenchmark\\": "benchmark"
         }
     },
     "suggest": {