--- a/phpbench.json
+++ b/phpbench.json
@@ -1,0 +1,6 @@
+{
+    "bootstrap": "vendor/autoload.php",
+    "extensions": [
+        "PhpBench\\Extensions\\XDebug\\XDebugExtension"
+    ]
+}