There are some performance issues which become very pronounced when:

- Reflecting large classes and their methods / properties (e.g. `__toString`), for example `PHPUnit_Framework_TestCase`.
- Calling `ReflectionClass#isSubClass` [many times](https://github.com/dantleech/phpactor/blob/master/lib/Complete/Provider/FetchProvider.php#L253).

The first performance issue can be greatly improved by caching in
`FindTypeFromAst` class, the second by caching calls to
`ReflectionClass#getInheritanceClassHierarchy`.

In this PR I included fixes to illustrate the issue, but ultimately it might be
better to take a different strategy (e.g. creating a service to resolve types
from AST).

In this PR I added a PHPBench benchmark which uses a copy of the `PHPUnit_Framework_TestCase` as a subject.

Before:

```
\Roave\BetterReflectionBenchmark\PhpunitFileBench

    reflect_phpunit_class         I0 P0         [μ Mo]/r: 0.547 0.547 (s)       [μSD μRSD]/r: 0.000s 0.00%
    reflection_to_string          I0 P0         [μ Mo]/r: 92.007 92.007 (s)     [μSD μRSD]/r: 0.000s 0.00%
```

After:

```
\Roave\BetterReflectionBenchmark\PhpunitFileBench

    reflect_phpunit_class         I0 P0         [μ Mo]/r: 0.479 0.479 (s)       [μSD μRSD]/r: 0.000s 0.00%
    reflection_to_string          I0 P0         [μ Mo]/r: 0.967 0.967 (s)       [μSD μRSD]/r: 0.000s 0.00%A
```