@Ocramius 2017-09-01
@dantleech can you check this on top of https://github.com/Roave/BetterReflection/pull/338 and see if it is still relevant?