@Ocramius 2017-09-02
See #346

Closing as `duplicate` here